+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 - Tweaks link added to admin menu
 - replaced user macro with {{user}} syntax (old syntax user#id is not supported anymore)
 - more formats for user macro and avatar support
-- rename list_users in members
+- rename list_users to members
+- rename list_projects to projects
 
 ### 0.5.7